channel, opening_control: Make first_blocknum u32
The `new_channel` constructor accepts u32, and the
`get_block_height` function returns u32, so the
extra 32 bits is unuseable anyway.

+++ lightningd/channel.h
@@ -92,7 +92,7 @@ struct channel {
 
 	/* Blockheight at creation, scans for funding confirmations
 	 * will start here */
-	u64 first_blocknum;
+	u32 first_blocknum;
 
 	/* Feerate range */
 	u32 min_possible_feerate, max_possible_feerate;
